Feature/exceptions details dto

+++ src/main/java/com/agh/surveys/service/user/UserService.java
@@ -1,5 +1,6 @@
 package com.agh.surveys.service.user;
 
+import com.agh.surveys.exception.BusinnessException;
 import com.agh.surveys.exception.user.UserExistsInDatabaseException;
 import com.agh.surveys.exception.user.UserNotFoundException;
 import com.agh.surveys.model.user.User;
@@ -7,6 +8,9 @@ import com.agh.surveys.model.user.dto.UserDto;
 import com.agh.surveys.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Collection;
+import java.util.stream.Stream;
 
 @Service
 public class UserService implements IUserService {
@@ -31,6 +35,13 @@ public class UserService implements IUserService {
 
     @Override
     public void removeUserByNick(String nick) {
+        User user = userRepository.getOne(nick);
+        if (!user.getManagedGroups().isEmpty()){
+            throw new BusinnessException("This User is a leader of a group and cannot be deleted!");
+        }
+        user.getUserGroups()
+                .forEach(group -> group.getGroupMembers().remove(user));
+
         userRepository.removeByUserNick(nick);
     }
